Basic support for galleries
users can now have a :gallery key in a point, looks like:

   :gallery: flickr,user=ripienaar,set=72157648567428473
   :type: gallery

this will render a custom type of point html in the marker popup that
has a link to the normal flickr flash gallery.  This is not ideal I want
it to behave that when clicked a modal style image browser just pops up
above the map and shows the images, closable with escape which puts you
right back on the map.  Failing to find a good existing library though
for this

Additionally types can be hidden on the type legend by setting
:legend: false

Templates for types of point popups can now be stored in different
erb files, right now only :gallery and :track types have their own erb
template which makes this a pre-requisite for #45

Relates to #47, #45